requirements.txt
plotly==5.11.0
tensorflow-cpu==2.11.0
scikit-learn==1.2.0
scipy==1.9.3
sklearn==0.0